added di & clean architecture?

+++ app/src/main/java/com/yelloyew/ewaweather/data/network/model/YandexParts.kt
@@ -2,7 +2,7 @@ package com.yelloyew.ewaweather.data.network.model
 
 data class YandexParts(
     val temp_avg: Int,
-    val cloudness: Int,
+    val cloudness: Float,
     val humidity: Int,
     val pressure_mm: Int
 )
